audit backend-tcop-fastpath: PASS (fix int16 zero-extension)
Independent function-by-function audit vs fastpath.c. One finding fixed:
numAFormats/nargs read pq_getmsgint(2) which C zero-extends into `int`
(0..65535); the port sign-extended (as i16 as i32), so a 0x8000-0xFFFF
count would go negative and skip the format-array alloc + the
nargs>FUNC_MAX_ARGS guard. Changed to `as i32` (zero-extend), matching C;
aformats[i]/rformat keep i16 (genuine int16), argsize keeps i32 (signed -1
NULL sentinel). audits/backend-tcop-fastpath.md added; CATALOG -> audited.

+++ crates/backend-tcop-fastpath/src/lib.rs
@@ -469,7 +469,9 @@ pub fn parse_fcall_arguments<'mcx>(
 
     // /* Get the argument format codes */
     // numAFormats = pq_getmsgint(msgBuf, 2);
-    numAFormats = pqformat::pq_getmsgint(msg_buf, 2)? as i16 as i32;
+    // C's pq_getmsgint(2) returns an `unsigned int` (the uint16 zero-extended),
+    // assigned to `int numAFormats` — so the value is 0..65535, never negative.
+    numAFormats = pqformat::pq_getmsgint(msg_buf, 2)? as i32;
     // if (numAFormats > 0)
     if numAFormats > 0 {
         // aformats = (int16 *) palloc(numAFormats * sizeof(int16));
@@ -484,7 +486,9 @@ pub fn parse_fcall_arguments<'mcx>(
     }
 
     // nargs = pq_getmsgint(msgBuf, 2);  /* # of arguments */
-    nargs = pqformat::pq_getmsgint(msg_buf, 2)? as i16 as i32;
+    // Zero-extended into C's `int nargs` (0..65535); the > FUNC_MAX_ARGS guard
+    // below rejects anything beyond 100.
+    nargs = pqformat::pq_getmsgint(msg_buf, 2)? as i32;
 
     // if (fip->flinfo.fn_nargs != nargs || nargs > FUNC_MAX_ARGS)
     if (fip.flinfo.fn_nargs as i32) != nargs || nargs > FUNC_MAX_ARGS as i32 {
